+++ lib/app.js
@@ -12,7 +12,7 @@ const BrowserTestRunner = require('./runners/browser_test_runner');
 const ProcessTestRunner = require('./runners/process_test_runner');
 const TapProcessTestRunner = require('./runners/tap_process_test_runner');
 const HookRunner = require('./runners/hook_runner');
-const cleanExit = require('./clean_exit');
+const cleanExit = require('./utils/clean_exit');
 const FileWatcher = require('./file_watcher');
 const LauncherFactory = require('./launcher-factory');
