+++ lib/browser_launcher.js
@@ -10,8 +10,8 @@ This file more or less figures out how to launch any browser on any platform.
 
 var Bluebird = require('bluebird');
 
-var fileutils = require('./fileutils');
-var envWithLocalPath = require('./env-with-local-path');
+var fileutils = require('./utils/fileutils');
+var envWithLocalPath = require('./utils/env-with-local-path');
 
 var executableExists = function(exe, config) {
   return fileutils.executableExists(exe, { env: envWithLocalPath(config) });
