+++ lib/config.js
@@ -21,10 +21,10 @@ const Bluebird = require('bluebird');
 
 const browser_launcher = require('./browser_launcher');
 const LauncherFactory = require('./launcher-factory');
-const Chars = require('./chars');
-const pad = require('./strutils').pad;
-const isa = require('./isa');
-const fileExists = require('./fileutils').fileExists;
+const Chars = require('./utils/chars');
+const pad = require('./utils/strutils').pad;
+const isa = require('./utils/isa');
+const fileExists = require('./utils/fileutils').fileExists;
 const uniqBy = require('lodash.uniqby');
 
 const knownBrowsers = require('./utils/known-browsers');
