+++ lib/launcher.js
@@ -5,7 +5,7 @@ const tmp = require('tmp');
 const path = require('path');
 const Bluebird = require('bluebird');
 
-const template = require('./strutils').template;
+const template = require('./utils/strutils').template;
 const ProcessCtl = require('./process-ctl');
 
 // setup graceful cleanup: removes the created directories when an uncaught exception occurs.
