+++ lib/process-ctl.js
@@ -8,8 +8,8 @@ const EventEmitter = require('events').EventEmitter;
 const spawnargs = require('spawn-args');
 const extend = require('lodash.assignin');
 
-const envWithLocalPath = require('./env-with-local-path');
-const fileutils = require('./fileutils');
+const envWithLocalPath = require('./utils/env-with-local-path');
+const fileutils = require('./utils/fileutils');
 const Process = require('./utils/process');
 
 const fileExists = fileutils.fileExists;
