+++ lib/reporters/dev/error_messages_panel.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const ScrollableTextPanel = require('./scrollable_text_panel');
-const Chars = require('../../chars');
+const Chars = require('../../utils/chars');
 
 module.exports = ScrollableTextPanel.extend({
   initialize(attrs) {
