+++ lib/reporters/dev/index.js
@@ -19,7 +19,7 @@ const constants = require('./constants');
 const RunnerTab = tabs.RunnerTab;
 const RunnerTabs = tabs.RunnerTabs;
 const Screen = require('./screen');
-const pad = require('../../strutils').pad;
+const pad = require('../../utils/strutils').pad;
 const ErrorMessagesPanel = require('./error_messages_panel');
 const Runner = require('./runner');
