+++ lib/reporters/dev/runner_tabs.js
@@ -12,8 +12,8 @@ When the tab is not selected, it hides the associated log panel.
 const SplitLogPanel = require('./split_log_panel');
 const View = require('./view');
 const Backbone = require('backbone');
-const pad = require('../../strutils').pad;
-const Chars = require('../../chars');
+const pad = require('../../utils/strutils').pad;
+const Chars = require('../../utils/chars');
 const Screen = require('./screen');
 const notifier = require('node-notifier');
 const constants = require('./constants');
