+++ lib/reporters/dev/scrollable_text_panel.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var View = require('./view');
-var splitLines = require('../../strutils').splitLines;
+var splitLines = require('../../utils/strutils').splitLines;
 var Screen = require('./screen');
 
 // This is a generic scrollable text viewer widget. Should be refactored
