+++ lib/reporters/dev/split_log_panel.js
@@ -4,8 +4,8 @@ var View = require('./view');
 var ScrollableTextPanel = require('./scrollable_text_panel');
 var tabs = require('./constants');
 var styledString = require('styled_string');
-var Chars = require('../../chars');
-var indent = require('../../strutils').indent;
+var Chars = require('../../utils/chars');
+var indent = require('../../utils/strutils').indent;
 var Screen = require('./screen');
 
 function failureDisplay(item) {
