+++ lib/reporters/dot_reporter.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const indent = require('../strutils').indent;
+const indent = require('../utils/strutils').indent;
 const printf = require('printf');
 
 module.exports = class DotReporter {
