+++ lib/reporters/tap_reporter.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const displayutils = require('../displayutils');
+const displayutils = require('../utils/displayutils');
 
 module.exports = class TapReporter {
   constructor(silent, out, config) {
