+++ lib/reporters/xunit_reporter.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var XmlDom = require('xmldom');
-var indent = require('../strutils').indent;
+var indent = require('../utils/strutils').indent;
 
 class XUnitReporter {
   constructor(silent, out, config) {
