+++ lib/runners/hook_runner.js
@@ -3,7 +3,7 @@
 const Bluebird = require('bluebird');
 const log = require('npmlog');
 
-const template = require('../strutils').template;
+const template = require('../utils/strutils').template;
 const ProcessCtl = require('../process-ctl');
 
 module.exports = class HookRunner {
