+++ lib/utils/add-to-PATH.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const cloneDeep = require('lodash.clonedeep');
-const isWin = require('./utils/is-win')();
+const isWin = require('./is-win')();
 let PATH = 'PATH';
 let delimiter = ':';
