+++ lib/utils/chars.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const isWin = require('./utils/is-win')();
+const isWin = require('./is-win')();
 
 // Special characters to use for drawing.
