+++ lib/utils/clean_exit.js
@@ -3,7 +3,7 @@
 module.exports = function cleanExit(code) {
   // Workaround for this node core bug <https://github.com/joyent/node/issues/3584>
   // Instead of using `process.exit(?code)`, use this instead.
-  //
+
   let draining = 0;
   function exit() {
     if (!(draining--)) {
