+++ lib/utils/displayutils.js
