+++ lib/utils/env-with-local-path.js
@@ -4,6 +4,7 @@ const path = require('path');
 const fs = require('fs');
 const addToPATH = require('./add-to-PATH');
 
+
 module.exports = function envWithLocalPath(config) {
   let configPath = path.join(config.cwd(), 'node_modules', '.bin');
   let modulesPath;
