+++ lib/utils/fileutils.js
@@ -7,7 +7,7 @@ const log = require('npmlog');
 
 const fsStatAsync = Bluebird.promisify(fs.stat);
 
-const isWin = require('./utils/is-win')();
+const isWin = require('./is-win')();
 
 exports.fileExists = function fileExists(path) {
   return fsStatAsync(path).then(stat => stat.isFile()).catchReturn(false);
