+++ lib/utils/reporter.js
@@ -4,7 +4,7 @@ const Bluebird = require('bluebird');
 const log = require('npmlog');
 
 const reporters = require('../reporters');
-const isa = require('../isa');
+const isa = require('./isa');
 const ReportFile = require('./report-file');
 
 function setupReporter(name, out, config, app) {
