+++ lib/utils/strutils.js
