+++ tests/ui/error_messages_panel_tests.js
@@ -3,7 +3,7 @@
 var expect = require('chai').expect;
 var screen = require('./fake_screen');
 var ErrorMessagesPanel = require('../../lib/reporters/dev/error_messages_panel');
-var Chars = require('../../lib/chars');
+var Chars = require('../../lib/utils/chars');
 
 var isWin = /^win/.test(process.platform);
